Backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
httpx==0.27.2
apscheduler
python-dotenv